requirements.txt
bs4
requests
google-cloud
google-cloud-scheduler
google-cloud-bigquery